+++ code/header/model/pvu_models/pv_converter_model/pvgu1.h
@@ -84,12 +84,12 @@ class PVGU1 : public PV_CONVERTER_MODEL
         void copy_from_const_model(const PVGU1& model);
 
         INTEGRAL_BLOCK active_current_commander;
-        double LVPL_active_power_change_rate;
+        float LVPL_active_power_change_rate;
         FIRST_ORDER_BLOCK LVPL_voltage_sensor;
         LVPL lvpl;
 
         FIRST_ORDER_BLOCK reactive_voltage_commander;
-        double HVRCR_voltage, HVRCR_current;
+        float HVRCR_voltage, HVRCR_current;
 };
 
 #endif // GENERATOR_MODEL_H
